todo-web-app/chatbot/backend/requirements-prod.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlmodel==0.0.16
pydantic==2.5.0
pydantic-settings==2.1.0
psycopg2-binary==2.9.9
alembic==1.13.1
openai==1.3.5
python-multipart==0.0.6
better-exceptions==0.3.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
uuid7==0.1.0